VCV-176: move hook, slim getCoverageCardStyle, total nomenclature
