+++ src/features/operations/components/operations-page-client.tsx
@@ -2,7 +2,7 @@
 
 import { Fragment, useEffect, useState } from 'react';
 import { useGetUserPermissions } from '@/api/user/hooks/use-get-user-permissions';
-import { useIsUgandaProgram } from '@/api/program/hooks/use-is-uganda-program';
+import { useIsUgandaProgram } from '@/lib/hooks/use-is-uganda-program';
 import PageShell from '@/components/layout/page-shell';
 import { Card, CardContent } from '@/components/ui/card';
 import { endOfMonth, format } from 'date-fns';
